pyproject.toml
[project]
name = "uk-elearning"
version = "0.1.0"
description = "UK eLearning Platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pythondi>=1.2.4",
    "redis>=6.2.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.35.0",
]
